Issue 388 Added db changes for v1.11.7 - 1.12.0 as a database patch

+++ database/migrations/classes/DoctrineMigrations/Version20130715003621RemoveObsoleteIndex.php
@@ -15,17 +15,32 @@ class Version20130715003621RemoveObsoleteIndex extends AbstractMigration
     private $tablePrefix = 'janus__';
 
     /**
-     * Remove obsolete remoteeid indexes which are no longer used
+     * Remove obsolete remoteeid indexes which are no longer used but were in the janus.sql for a while but was never
+     * part of any upgrade instructions and was never useful
+     *
      * @param Schema $schema
      */
     public function up(Schema $schema)
     {
-        $this->addSql("
+        $this->removeObsoleteIndex($schema, 'allowedEntity');
+        $this->removeObsoleteIndex($schema, 'blockedEntity');
+    }
+
+    /**
+     * @param Schema $schema
+     * @param string $tableName
+     */
+    private function removeObsoleteIndex(Schema $schema, $tableName)
+    {
+        $prefixedTableName = $this->tablePrefix . $tableName;
+        if ($schema->getTable($prefixedTableName)->hasIndex('remoteeid')) {
+            $this->addSql("
             DROP INDEX remoteeid
-                ON janus__allowedEntity");
-        $this->addSql("
-            DROP INDEX remoteeid
-                ON janus__blockedEntity");
+              ON janus__allowedEntity");
+            $this->addSql("
+              DROP INDEX remoteeid
+              ON janus__blockedEntity");
+        }
     }
 
     /**
@@ -33,11 +48,6 @@ class Version20130715003621RemoveObsoleteIndex extends AbstractMigration
      */
     public function down(Schema $schema)
     {
-        $this->addSql("
-          CREATE INDEX remoteeid
-              ON janus__allowedEntity (remoteeid)");
-        $this->addSql("
-          CREATE INDEX remoteeid
-              ON janus__blockedEntity (remoteeid)");
+        //
     }
 }
